Address `bs-` and `bsc-` warnings
The warning message:
```
The field 'bs-dependencies' found in the package config of 'vite-react-rescript-starter' is deprecated and will be removed in a future version.
Use 'dependencies' instead.

The field 'bs-dev-dependencies' found in the package config of 'vite-react-rescript-starter' is deprecated and will be removed in a future version.
Use 'dev-dependencies' instead.

The field 'bsc-flags' found in the package config of 'vite-react-rescript-starter' is deprecated and will be removed in a future version.
Use 'compiler-flags' instead.
```